Move Context into Core
