+++ nanoc-core/nanoc-core.manifest
@@ -7,6 +7,7 @@ lib/nanoc/core/version.rb
 
 lib/nanoc/core/binary_content.rb
 lib/nanoc/core/content.rb
+lib/nanoc/core/context.rb
 lib/nanoc/core/contracts_support.rb
 lib/nanoc/core/core_ext/array.rb
 lib/nanoc/core/core_ext/hash.rb
