+++ nanoc/lib/nanoc/base/entities/code_snippet.rb
@@ -34,7 +34,7 @@ module Nanoc::Int
     #
     # @return [void]
     def load
-      eval('def self.use_helper(mod); Nanoc::Int::Context.instance_eval { include mod }; end', TOPLEVEL_BINDING)
+      eval('def self.use_helper(mod); Nanoc::Core::Context.instance_eval { include mod }; end', TOPLEVEL_BINDING)
       eval(@data, TOPLEVEL_BINDING, @filename)
       nil
     end
