+++ nanoc/lib/nanoc/filters/erubis.rb
@@ -15,7 +15,7 @@ module Nanoc::Filters
     # @return [String] The filtered content
     def run(content, _params = {})
       # Create context
-      context = ::Nanoc::Int::Context.new(assigns)
+      context = ::Nanoc::Core::Context.new(assigns)
 
       # Get binding
       proc = assigns[:content] ? -> { assigns[:content] } : nil
