+++ nanoc/nanoc.manifest
@@ -14,7 +14,6 @@ lib/nanoc/base/entities/checksum_collection.rb
 lib/nanoc/base/entities/code_snippet.rb
 lib/nanoc/base/entities/configuration.rb
 lib/nanoc/base/entities/configuration-schema.json
-lib/nanoc/base/entities/context.rb
 lib/nanoc/base/entities/dependency.rb
 lib/nanoc/base/entities/directed_graph.rb
 lib/nanoc/base/entities/document.rb
